requirements.txt
certifi==2020.12.5
chardet==3.0.4
graphviz==0.8.4
idna==2.6
mxnet==1.7.0.post1
numpy==1.16.6
Pillow==8.0.1
requests==2.18.4
urllib3==1.22